% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiPlusZ.R
\docType{package}
\name{chiPlusZ-package}
\alias{chiPlusZ}
\alias{chiPlusZ-package}
\title{chiPlusZ: Chi-Squared + SPSS-Style Column Proportion Z-Tests}
\description{
Replicates SPSS's "Compare Column Proportions" workflow for two-way
contingency tables.  The main entry point is \code{\link[=crosstabs]{crosstabs()}}.
}
\author{
\strong{Maintainer}: Gabija Vaitkevičiūtė \email{info@biomatika.lt}

Authors:
\itemize{
  \item Gabija Vaitkevičiūtė \email{info@biomatika.lt}
}

}
\keyword{internal}
