% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiPlusZ.R
\name{print.crosstabs}
\alias{print.crosstabs}
\title{Print method for \code{crosstabs} objects}
\usage{
\method{print}{crosstabs}(x, ...)
}
\arguments{
\item{x}{A \code{crosstabs} object returned by \code{\link[=crosstabs]{crosstabs()}}.}

\item{...}{Currently unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for \code{crosstabs} objects
}
